requirements.txt
# Core dependencies
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
accelerate>=0.24.0
trl>=0.7.0

# Audio processing
librosa>=0.10.0
soundfile>=0.12.0

# API clients
openai>=1.0.0
google-generativeai>=0.3.0
anthropic>=0.7.0

# PEFT and optimization
peft>=0.6.0
deepspeed>=0.12.0
bitsandbytes>=0.41.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Visualization and logging
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.14.0
wandb>=0.15.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
requests>=2.31.0
aiohttp>=3.8.0

# Development tools
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
mypy>=1.5.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
